Divide code to separate files
